migrations and controllers
created migration files for question and response. Changed the retrieve function in teams and user controllers to return the questions when teams/:id is called and return responses when user/:token is called.
TODO: Find a way to return responses when calling questions and find a way to return question when calling responses.